Extract state and timing to wrapper component
